requirement.txt
streamlit
yfinance
requests
pandas
bs4
openai
python-dotenv